backend/pyproject.toml
[project]
name = 'book-inventory-backend'
version = '0.1.0'
description = 'Local-first FastAPI backend for physical book inventory.'
requires-python = '>=3.10'
dependencies = [
    'fastapi>=0.111.0',
    'uvicorn[standard]>=0.27.0',
    'sqlmodel>=0.0.16',
    'aiosqlite>=0.20.0',
    'alembic>=1.13.1',
    'pydantic-settings>=2.2.1',
    'httpx>=0.27.0',
    'python-multipart>=0.0.7',
    'python-jose[cryptography]>=3.3.0',
    'passlib[bcrypt]>=1.7.4',
    'bcrypt>=4.0.0,<5.0.0'
]

[project.optional-dependencies]
dev = [
    'pytest>=8.0.0',
    'pytest-asyncio>=0.23.0',
    'ruff>=0.4.0',
    'mypy>=1.10.0'
]

[build-system]
requires = ['setuptools>=64', 'wheel']
build-backend = 'setuptools.build_meta'

[tool.setuptools]
package-dir = {"" = "app"}

[tool.setuptools.packages.find]
where = ["app"]
